major refactoring to support `views[].metadata.appConfiguration` ...
* updated to `mmif-python==1.0.11`
* parameter type casting now happens in `ClamsApp` class, instead of on the restifier wrapper layer (updated all test cases for this)
* all "raw" parameters passed to the app are now assumed to be a list of string values, directly from `flask.request.args.to_dict(flat=False)`
