+++ requirements.txt
@@ -1,4 +1,4 @@
-mmif-python==1.0.10
+mmif-python==1.0.11
 
 Flask>=2
 Flask-RESTful>=0.3.9
